dev: extract clock to new package
