+++ cSpell.json
@@ -34,6 +34,7 @@
         "completei",
         "connectionless",
         "Containerfile",
+        "conv",
         "curr",
         "Cyberneering",
         "dashmap",
@@ -116,6 +117,7 @@
         "rngs",
         "rosegment",
         "routable",
+        "rstest",
         "rusqlite",
         "RUSTDOCFLAGS",
         "RUSTFLAGS",
