+++ packages/clock/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+description = "A library to a clock for the torrust tracker."
+keywords = ["library", "clock", "torrents"]
+name = "torrust-tracker-clock"
+readme = "README.md"
+
+authors.workspace = true
+categories.workspace = true
+documentation.workspace = true
+edition.workspace = true
+homepage.workspace = true
+license.workspace = true
+publish.workspace = true
+repository.workspace = true
+rust-version.workspace = true
+version.workspace = true
+
+[dependencies]
+lazy_static = "1"
+chrono = { version = "0", default-features = false, features = ["clock"] }
+
+torrust-tracker-primitives = { version = "3.0.0-alpha.12-develop", path = "../primitives" }
+
+[dev-dependencies]
