+++ packages/clock/README.md
@@ -1,0 +1,11 @@
+# Torrust Tracker Clock
+
+A library to provide a working and mockable clock for the [Torrust Tracker](https://github.com/torrust/torrust-tracker).
+
+## Documentation
+
+[Crate documentation](https://docs.rs/torrust-tracker-torrent-clock).
+
+## License
+
+The project is licensed under the terms of the [GNU AFFERO GENERAL PUBLIC LICENSE](./LICENSE).
