+++ packages/clock/tests/integration.rs
@@ -4,17 +4,16 @@
 //! cargo test --test integration
 //! ```
 
-use torrust_tracker_clock::clock;
-mod common;
-mod servers;
+//mod common;
+mod clock;
 
 /// This code needs to be copied into each crate.
 /// Working version, for production.
 #[cfg(not(test))]
 #[allow(dead_code)]
-pub(crate) type CurrentClock = clock::Working;
+pub(crate) type CurrentClock = torrust_tracker_clock::clock::Working;
 
 /// Stopped version, for testing.
 #[cfg(test)]
 #[allow(dead_code)]
-pub(crate) type CurrentClock = clock::Stopped;
+pub(crate) type CurrentClock = torrust_tracker_clock::clock::Stopped;
