+++ packages/torrent-repository/src/lib.rs
@@ -1,5 +1,7 @@
 use std::sync::Arc;
 
+use torrust_tracker_clock::clock;
+
 pub mod entry;
 pub mod repository;
 
@@ -13,3 +15,14 @@ pub type TorrentsRwLockStdMutexTokio = repository::RwLockStd<EntryMutexTokio>;
 pub type TorrentsRwLockTokio = repository::RwLockTokio<EntrySingle>;
 pub type TorrentsRwLockTokioMutexStd = repository::RwLockTokio<EntryMutexStd>;
 pub type TorrentsRwLockTokioMutexTokio = repository::RwLockTokio<EntryMutexTokio>;
+
+/// This code needs to be copied into each crate.
+/// Working version, for production.
+#[cfg(not(test))]
+#[allow(dead_code)]
+pub(crate) type CurrentClock = clock::Working;
+
+/// Stopped version, for testing.
+#[cfg(test)]
+#[allow(dead_code)]
+pub(crate) type CurrentClock = clock::Stopped;
