+++ src/bootstrap/app.rs
@@ -13,13 +13,13 @@
 //! 4. Initialize the domain tracker.
 use std::sync::Arc;
 
+use torrust_tracker_clock::static_time;
 use torrust_tracker_configuration::Configuration;
 
 use super::config::initialize_configuration;
 use crate::bootstrap;
 use crate::core::services::tracker_factory;
 use crate::core::Tracker;
-use crate::shared::clock::static_time;
 use crate::shared::crypto::ephemeral_instance_keys;
 
 /// It loads the configuration from the environment and builds the main domain [`Tracker`] struct.
