+++ src/shared/mod.rs
@@ -1,8 +1,6 @@
 //! Modules with generic logic used by several modules.
 //!
 //! - [`bit_torrent`]: `BitTorrent` protocol related logic.
-//! - [`clock`]: Times services.
 //! - [`crypto`]: Encryption related logic.
 pub mod bit_torrent;
-pub mod clock;
 pub mod crypto;
