+++ tests/common/mod.rs
@@ -1,3 +1,4 @@
+pub mod clock;
 pub mod fixtures;
 pub mod http;
 pub mod udp;
